+++ app/application.py
@@ -12,6 +12,19 @@ from utils.temp import TempFile
 
 from .routes import router as ImplRouter
 
+DESCRIPTION = """
+**An alternative implement of Imjad API**
+
+- *Documents*:
+    - [Redoc](/docs) (Easier to read and more beautiful)
+    - [Swagger UI](/docs/test) (Integrated interactive testing function)
+
+Project: [mixmoe/HibiAPI](https://github.com/mixmoe/HibiAPI)
+
+![](https://img.shields.io/github/stars/mixmoe/HibiAPI?color=brightgreen&logo=github&style=for-the-badge)
+"""
+
+
 if Config["log"]["sentry"]["enabled"].as_bool():
     sentry_sdk.init(
         dsn=Config["log"]["sentry"]["dsn"].as_str(),
@@ -25,7 +38,7 @@ app = FastAPI(
     debug=Config["debug"].as_bool(),
     title="HibiAPI",
     version=Config["version"].as_str(),
-    description="An alternative implement of Imjad API",
+    description=DESCRIPTION,
     docs_url="/docs/test",
     redoc_url="/docs",
 )
