+++ utils/net.py
@@ -1,38 +1,26 @@
 import asyncio
-import inspect
-from enum import Enum
-from fnmatch import fnmatch
+from functools import wraps
 from threading import current_thread
 from types import TracebackType
-from typing import Any, Callable, Dict, List, Mapping, Optional, Tuple, Type, Union
-from urllib.parse import ParseResult, urlparse
+from typing import Any, Callable, Coroutine, Dict, Optional, Type, TypeVar, Union
 
-from fastapi.routing import APIRouter
 from httpx import (
     URL,
     AsyncClient,
     Cookies,
+    HTTPError,
+    HTTPStatusError,
     Request,
     Response,
     ResponseNotRead,
     TransportError,
 )
-from pydantic import AnyHttpUrl, validate_arguments
-from pydantic.errors import UrlHostError
 
-from .decorators import Retry
+from .decorators import Retry, TimeIt
+from .exceptions import UpstreamAPIException
 from .log import logger
 
-
-def exclude_params(func: Callable, params: Mapping[str, Any]) -> Dict[str, Any]:
-    func_params = inspect.signature(func).parameters
-    return {k: v for k, v in params.items() if k in func_params}
-
-
-class SlashRouter(APIRouter):
-    def api_route(self, path: str, **kwargs):
-        path = path if path.startswith("/") else ("/" + path)
-        return super().api_route(path, **kwargs)
+_AsyncCallable = TypeVar("_AsyncCallable", bound=Callable[..., Coroutine])
 
 
 class AsyncHTTPClient(AsyncClient):
@@ -117,60 +105,16 @@ class BaseNetClient:
         )
 
 
-class BaseEndpoint:
-    type_checking: bool = True
+def catch_network_error(function: _AsyncCallable) -> _AsyncCallable:
+    function = TimeIt(function)
 
-    def __init__(self, client: AsyncHTTPClient):
-        self.client = client
+    @wraps(function)
+    async def wrapper(*args, **kwargs):
+        try:
+            return await function(*args, **kwargs)
+        except HTTPStatusError as e:
+            raise UpstreamAPIException(detail=e.response.text) from e
+        except HTTPError as e:
+            raise UpstreamAPIException from e
 
-    @staticmethod
-    def _join(base: str, endpoint: str, params: Dict[str, Any]) -> URL:
-        host: ParseResult = urlparse(base)
-        params = {
-            k: (v.value if isinstance(v, Enum) else v)
-            for k, v in params.items()
-            if v is not None
-        }
-        return URL(
-            url=ParseResult(
-                scheme=host.scheme,
-                netloc=host.netloc,
-                path=endpoint.format(**params),
-                params="",
-                query="",
-                fragment="",
-            ).geturl(),
-            params=params,
-        )
-
-    def __getattribute__(self, name: str) -> Any:
-        obj = super().__getattribute__(name)
-        if name.startswith("_"):
-            return obj
-        elif not callable(obj):
-            return obj
-        elif not self.type_checking:
-            return obj
-        return validate_arguments(obj)
-
-
-class BaseHostUrl(AnyHttpUrl):
-    allowed_hosts: List[str] = []
-
-    @classmethod
-    def validate_host(
-        cls, parts: Dict[str, str]
-    ) -> Tuple[str, Optional[str], str, bool]:
-        host, tld, host_type, rebuild = super().validate_host(parts)
-        if not cls._check_domain(host):
-            raise UrlHostError(allowed=cls.allowed_hosts)
-        return host, tld, host_type, rebuild
-
-    @classmethod
-    def _check_domain(cls, host: str) -> bool:
-        return any(
-            filter(
-                lambda x: fnmatch(host, x),  # type:ignore
-                cls.allowed_hosts,
-            )
-        )
+    return wrapper  # type:ignore
